requirements.txt
pathway
sentence-transformers
numpy
httpx
requests
python-dotenv
fastapi
uvicorn
streamlit
requests
